Added PrivateToPublicKeyDetailsFragmentInIsolationTest. Refactored code.| #1746
